+++ containers/tests/helpers.py
@@ -1,9 +1,17 @@
 """Helpers for the container tests."""
 import uuid
+import dateutil.parser
 
+from django.utils import dateformat
 from test_plus.test import TestCase
 
-from containers.models import Container, STATE_INITIAL
+from containers.models import (
+    Container,
+    STATE_INITIAL,
+    STATE_CREATED,
+    STATE_RUNNING,
+    STATE_EXITED,
+)
 from containers.tests.factories import ProjectFactory, ContainerFactory
 
 
@@ -43,11 +51,35 @@ class TestBase(TestCase):
         self.fake_uuid = uuid.uuid4()
 
 
+def log_entry1():
+    """First log entry."""
+    dt = dateutil.parser.parse("2021-01-01 01:01:01.000001+00:00")
+    return dt, "{} 2021/01/01 10:00:00 [info] Log entry 1".format(
+        dateformat.format(dt.replace(tzinfo=None), "c") + "000Z"
+    )
+
+
+def log_entry2():
+    """Second log entry, same second but different millisecond."""
+    dt = dateutil.parser.parse("2021-01-01 01:01:01.500001+00:00")
+    return dt, "{} 2021/01/01 10:00:00 [info] Log entry 2".format(
+        dateformat.format(dt.replace(tzinfo=None), "c") + "000Z"
+    )
+
+
+def log_entry3():
+    """Third log entry happening the next second"""
+    dt = dateutil.parser.parse("2021-01-01 01:01:02.000001+00:00")
+    return dt, "{} 2021/01/01 10:00:01 [info] Log entry 3".format(
+        dateformat.format(dt.replace(tzinfo=None), "c") + "000Z"
+    )
+
+
 class DockerMock:
     """Class to mock calls to Docker API."""
 
     @classmethod
-    def pull(self, repository, tag, stream, decode):
+    def pull(cls, repository, tag, stream, decode):
         return [
             {
                 "progressDetail": {"total": "total", "current": "current"},
@@ -56,23 +88,41 @@ class DockerMock:
         ]
 
     @classmethod
-    def inspect_image(self, image):
+    def inspect_image(cls, image):
         return {"Id": "1"}
 
     @classmethod
+    def inspect_container_started(cls, container):
+        return {"State": {"Status": STATE_RUNNING}}
+
+    @classmethod
+    def inspect_container_stopped(cls, container):
+        return {"State": {"Status": STATE_EXITED}}
+
+    @classmethod
     def create_container(
-        self, detach, image, environment, command, ports, host_config
+        cls, detach, image, environment, command, ports, host_config
     ):
-        return {"Id": "9"}
+        return {"Id": "9", "State": {"Status": STATE_CREATED}}
 
     @classmethod
-    def create_host_config(self, port_bindings, ulimits):
+    def create_host_config(cls, port_bindings, ulimits):
         return
 
     @classmethod
-    def start(self, container):
+    def start(cls, container):
         return
 
     @classmethod
-    def stop(self, container):
+    def stop(cls, container):
         return
+
+    @classmethod
+    def logs(cls, container, timestamps):
+        return "\n".join(
+            [log_entry1()[1], log_entry2()[1], log_entry3()[1]]
+        ).encode("utf-8")
+
+    @classmethod
+    def logs_since(cls, container, timestamps, since):
+        return "\n".join([log_entry2()[1], log_entry3()[1]]).encode("utf-8")