+++ containers/tests/test_models.py
@@ -1,8 +1,10 @@
 """Tests for the container models"""
 import json
+from datetime import timedelta
 
 from django.forms import model_to_dict
 from django.urls import reverse
+from django.utils import timezone
 from django.utils.timezone import localtime
 
 from containers.models import (
@@ -10,6 +12,8 @@ from containers.models import (
     STATE_INITIAL,
     LOG_LEVEL_INFO,
     ContainerLogEntry,
+    PROCESS_DOCKER,
+    PROCESS_OBJECT,
 )
 from containers.tests.factories import ContainerLogEntryFactory
 from containers.tests.helpers import TestBase
@@ -138,7 +142,28 @@ class TestContainerLogEntry(TestBase):
         self.log_entry_no_user = ContainerLogEntryFactory(
             container=self.container1
         )
+        self.log_entry_docker1 = ContainerLogEntryFactory(
+            container=self.container1,
+            user=self.superuser,
+            process=PROCESS_DOCKER,
+            date_docker_log=timezone.now() - timedelta(minutes=2),
+        )
+        self.log_entry_docker2 = ContainerLogEntryFactory(
+            container=self.container1,
+            user=self.superuser,
+            process=PROCESS_DOCKER,
+            date_docker_log=timezone.now() - timedelta(minutes=2),
+        )
         self.data = {
+            "process": PROCESS_OBJECT,
+            "level": LOG_LEVEL_INFO,
+            "text": "Log entry",
+            "user": self.superuser,
+            "container": self.container1,
+        }
+        self.data_docker_log = {
+            "process": PROCESS_DOCKER,
+            "date_docker_log": timezone.now() - timedelta(minutes=2),
             "level": LOG_LEVEL_INFO,
             "text": "Log entry",
             "user": self.superuser,
@@ -152,6 +177,18 @@ class TestContainerLogEntry(TestBase):
             "user": self.superuser.pk,
             "container": self.container1.pk,
             "id": log_entry.pk,
+            "date_docker_log": None,
+        }
+        self.assertEqual(model_to_dict(log_entry), expected)
+
+    def test_initialization_docker_log(self):
+        log_entry = ContainerLogEntry.objects.create(**self.data_docker_log)
+        expected = {
+            **self.data_docker_log,
+            "user": self.superuser.pk,
+            "container": self.container1.pk,
+            "id": log_entry.pk,
+            "date_docker_log": log_entry.date_docker_log,
         }
         self.assertEqual(model_to_dict(log_entry), expected)
 
@@ -163,23 +200,81 @@ class TestContainerLogEntry(TestBase):
             ),
         )
 
+    def test_get_date_docker_log(self):
+        self.assertEqual(
+            self.log_entry_docker1.get_date_docker_log(),
+            localtime(self.log_entry_docker1.date_docker_log).strftime(
+                "%Y-%m-%d %H:%M:%S"
+            ),
+        )
+
+    def test_get_date_docker_log_none(self):
+        self.assertEqual(self.log_entry.get_date_docker_log(), "")
+
+    def test_get_date_order_by(self):
+        self.assertEqual(
+            self.log_entry.get_date_order_by(),
+            localtime(self.log_entry.date_created).strftime(
+                "%Y-%m-%d %H:%M:%S"
+            ),
+        )
+
+    def test_get_date_order_by_docker_log(self):
+        self.assertEqual(
+            self.log_entry_docker1.get_date_order_by(),
+            localtime(self.log_entry_docker1.date_docker_log).strftime(
+                "%Y-%m-%d %H:%M:%S"
+            ),
+        )
+
     def test___str__(self):
         self.assertEqual(
             str(self.log_entry),
-            "[{} {} {}] {}".format(
-                self.log_entry.get_date_created(),
+            "[{} {} {}] ({}) {}".format(
+                self.log_entry.get_date_order_by(),
                 self.log_entry.level.upper(),
                 self.log_entry.user.username,
+                self.log_entry.process.capitalize(),
                 self.log_entry.text,
+            ),
+        )
+
+    def test___str___docker_log(self):
+        self.assertEqual(
+            str(self.log_entry_docker1),
+            "[{} {} {}] ({}) {}".format(
+                self.log_entry_docker1.get_date_order_by(),
+                self.log_entry_docker1.level.upper(),
+                self.log_entry_docker1.user.username,
+                self.log_entry_docker1.process.capitalize(),
+                self.log_entry_docker1.text,
             ),
         )
 
     def test___str___no_user(self):
         self.assertEqual(
             str(self.log_entry_no_user),
-            "[{} {} anonymous] {}".format(
-                self.log_entry_no_user.get_date_created(),
+            "[{} {} anonymous] ({}) {}".format(
+                self.log_entry_no_user.get_date_order_by(),
                 self.log_entry_no_user.level.upper(),
+                self.log_entry_no_user.process.capitalize(),
                 self.log_entry_no_user.text,
             ),
+        )
+
+    def test_containerlogentrymanager_merge_order(self):
+        self.assertEqual(
+            [
+                self.log_entry_docker1,
+                self.log_entry_docker2,
+                self.log_entry,
+                self.log_entry_no_user,
+            ],
+            list(ContainerLogEntry.objects.merge_order()),
+        )
+
+    def test_containerlogentrymanager_get_date_last_docker_log(self):
+        self.assertEqual(
+            self.log_entry_docker2.get_date_docker_log(),
+            ContainerLogEntry.objects.get_date_last_docker_log(),
         )